ajuste no arquivo de editar.php

+++ index.php
@@ -26,11 +26,9 @@ if ($sql->execute()) { // Verifica se a consulta foi executada com sucesso
     <title>Sua Página PHP com Bootstrap</title>
 
     <!-- Incluindo os arquivos CSS do Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <!-- Incluindo os arquivos JavaScript do Bootstrap (opcional, dependendo das funcionalidades que você planeja usar) -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-</head>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script></head>
 
 <body>
     <!-- Conteúdo da sua página PHP aqui -->
@@ -58,7 +56,7 @@ if ($sql->execute()) { // Verifica se a consulta foi executada com sucesso
                     <td><?php echo $usuario['email']; ?></td>
                     <td>
                         <a href="editar.php?id=<?php echo $usuario['id']; ?>">[ Editar ]</a>
-                        <a href="excluir.php?id=<?php echo $usuario['id']; ?>">[ Excluir ]</a>
+                        <a href="excluir.php?id=<?php echo $usuario['id']; ?>" onclick="return confirm('Tem certeza que deseja excluir?')">[ Excluir ]</a>
                     </td>
                 </tr>
             <?php endforeach; ?>
